issues fixed index.html
fixed an issue that caused the "predicting for...." element to glitch out.

Improved the snowflake animation, including higher contrast in light mode and randomized starting position to avoid the "clumping" of snowflakes at the top of the page upon loading for the first time. 

Changed the style of the snow day prediction element moving from a 'rounded line' bar style to a 'rounded rectangle' with shadows that are more visually consistent to the other design elements.